+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/list/MessageListController.kt
@@ -629,6 +629,13 @@ public class MessageListController(
     }
 
     /**
+     * Hides the unread label in the messages list (if already visible).
+     */
+    public fun hideUnreadSeparator() {
+        unreadLabelState.value = null
+    }
+
+    /**
      * Scrolls to the first unread message in the channel.
      */
     public fun scrollToFirstUnreadMessage() {